refactor: extract conversion to own file
